+++ src/gevent/_socket3.py
@@ -654,9 +654,48 @@ if hasattr(_socket, "socketpair"):
         a = socket(family, type, proto, a.detach())
         b = socket(family, type, proto, b.detach())
         return a, b
+elif sys.version_info[:2] >= (3, 6):
+    # Origin: https://gist.github.com/4325783, by Geert Jansen.  Public domain.
+    # gevent: taken from 3.6 release. Expected to be used only on Win/3.6
+    def socketpair(family=AF_INET, type=SOCK_STREAM, proto=0):
+        if family == AF_INET:
+            host = _LOCALHOST
+        elif family == AF_INET6:
+            host = _LOCALHOST_V6
+        else:
+            raise ValueError("Only AF_INET and AF_INET6 socket address families "
+                             "are supported")
+        if type != SOCK_STREAM:
+            raise ValueError("Only SOCK_STREAM socket type is supported")
+        if proto != 0:
+            raise ValueError("Only protocol zero is supported")
+
+        # We create a connected TCP socket. Note the trick with
+        # setblocking(False) that prevents us from having to create a thread.
+        lsock = socket(family, type, proto)
+        try:
+            lsock.bind((host, 0))
+            lsock.listen()
+            # On IPv6, ignore flow_info and scope_id
+            addr, port = lsock.getsockname()[:2]
+            csock = socket(family, type, proto)
+            try:
+                csock.setblocking(False)
+                try:
+                    csock.connect((addr, port))
+                except (BlockingIOError, InterruptedError):
+                    pass
+                csock.setblocking(True)
+                ssock, _ = lsock.accept()
+            except:
+                csock.close()
+                raise
+        finally:
+            lsock.close()
+        return (ssock, csock)
 
 elif 'socketpair' in __implements__:
-    # Win32: not available
+    # Win32: not available prior to 3.6
     # Multiple imports can cause this to be missing if _socketcommon
     # was successfully imported, leading to subsequent imports to cause
     # ValueError