+++ src/gevent/subprocess.py
@@ -141,6 +141,11 @@ if sys.version_info[:2] >= (3, 5):
     except:
         MAXFD = 256
 
+if sys.version_info[:2] >= (3, 6):
+    # This was added to __all__ for windows in 3.6
+    __extra__.remove('STARTUPINFO')
+    __implements__.append('STARTUPINFO')
+
 actually_imported = copy_globals(__subprocess__, globals(),
                                  only_names=__imports__,
                                  ignore_missing_names=True)